split up rhobot main cli from utility functions
